+++ src/Splitio/Services/Cache/Interfaces/ISegmentCache.cs
@@ -1,18 +1,7 @@
-﻿using System.Collections.Generic;
-
-namespace Splitio.Services.Cache.Interfaces
+﻿namespace Splitio.Services.Cache.Interfaces
 {
-    public interface ISegmentCache : ISegmentCacheAsync
+    public interface ISegmentCache : ISegmentCacheConsumer, ISegmentCacheProducer
     {
-        void AddToSegment(string segmentName, List<string> segmentKeys);
-        void RemoveFromSegment(string segmentName, List<string> segmentKeys);
-        bool IsInSegment(string segmentName, string key);
-        void SetChangeNumber(string segmentName, long changeNumber);
-        long GetChangeNumber(string segmentName);
-        void Clear();
-        List<string> GetSegmentNames();
-        List<string> GetSegmentKeys(string segmentName);
-        int SegmentsCount();
-        int SegmentKeysCount();
+        
     }
 }
