+++ src/Splitio/Services/Client/Classes/SplitClient.cs
@@ -43,7 +43,6 @@ namespace Splitio.Services.Client.Classes
         protected ISplitManager _manager;
         protected IEventsLog _eventsLog;
         protected ITrafficTypeValidator _trafficTypeValidator;
-        protected ISegmentCache _segmentCache;
         protected IBlockUntilReadyService _blockUntilReadyService;
         protected IFactoryInstantiationsService _factoryInstantiationsService;
         protected ISplitParser _splitParser;
